backend/requirements.txt
fastapi
uvicorn
pydantic
weasyprint
jinja2
openai
pypdf
python-multipart
python-dotenv
gunicorn
uvicorn